Add ClientIDAPI_UnitTest file & move relevent tests over from API_UnitTest file
