+++ CorePayments/src/test/java/com/paypal/android/corepayments/APIUnitTest.kt
@@ -45,8 +45,7 @@ class APIUnitTest {
 
     @Before
     fun beforeEach() {
-        sut = API(configuration, http, httpRequestFactory, analyticsService)
-        API.clientIDCache.evictAll()
+        sut = API(configuration, http, httpRequestFactory)
     }
 
     @Test
@@ -60,135 +59,5 @@ class APIUnitTest {
 
         val result = sut.send(apiRequest)
         assertSame(httpResponse, result)
-    }
-
-    @Test
-    fun `fetchCachedOrRemoteClientID() sends oauth api request when value not in cache`() = runTest {
-        val apiRequestSlot = slot<APIRequest>()
-        every {
-            httpRequestFactory.createHttpRequestFromAPIRequest(
-                capture(apiRequestSlot),
-                configuration
-            )
-        } returns httpRequest
-
-        coEvery { http.send(httpRequest) } returns clientIdSuccessResponse
-
-        sut.fetchCachedOrRemoteClientID()
-
-        val apiRequest = apiRequestSlot.captured
-        assertEquals(HttpMethod.GET, apiRequest.method)
-        assertEquals("v1/oauth2/token", apiRequest.path)
-    }
-
-    @Test
-    fun `fetchCachedOrRemoteClientID() puts value in cache after fetched`() = runTest {
-        val apiRequestSlot = slot<APIRequest>()
-        every {
-            httpRequestFactory.createHttpRequestFromAPIRequest(
-                capture(apiRequestSlot),
-                configuration
-            )
-        } returns httpRequest
-
-        coEvery { http.send(httpRequest) } returns clientIdSuccessResponse
-
-        sut.fetchCachedOrRemoteClientID()
-
-        assertEquals(API.clientIDCache.get("fake-access-token"), "sample-client-id")
-    }
-
-    @Test
-    fun `fetchCachedOrRemoteClientID() returns cached value when exists in cache`() = runTest {
-        API.clientIDCache.put("fake-access-token", "cached-id-123")
-
-        val clientID = sut.fetchCachedOrRemoteClientID()
-        assertEquals(clientID, "cached-id-123")
-    }
-
-    @Test
-    fun `fetchCachedOrRemoteClientID() returns client id from JSON`() = runTest {
-        every {
-            httpRequestFactory.createHttpRequestFromAPIRequest(any(), any())
-        } returns httpRequest
-
-        coEvery { http.send(httpRequest) } returns clientIdSuccessResponse
-
-        val result = sut.fetchCachedOrRemoteClientID()
-        assertEquals("sample-client-id", result)
-    }
-
-    @Test
-    fun `fetchCachedOrRemoteClientID() throws no response data error when http response has no body`() =
-        runTest {
-
-            every {
-                httpRequestFactory.createHttpRequestFromAPIRequest(any(), any())
-            } returns httpRequest
-
-            val noBodyHttpResponse = HttpResponse(200, httpResponseHeaders)
-            coEvery { http.send(httpRequest) } returns noBodyHttpResponse
-
-            var capturedError: PayPalSDKError? = null
-            try {
-                sut.fetchCachedOrRemoteClientID()
-            } catch (e: PayPalSDKError) {
-                capturedError = e
-            }
-            assertEquals(Code.NO_RESPONSE_DATA.ordinal, capturedError?.code)
-            assertEquals("sample-correlation-id", capturedError?.correlationID)
-        }
-
-    @Test
-    fun `fetchCachedOrRemoteClientID() throws data parsing error when http response is missing client id`() =
-        runTest {
-
-            every {
-                httpRequestFactory.createHttpRequestFromAPIRequest(any(), any())
-            } returns httpRequest
-
-            val httpResponseWithoutClientId =
-                HttpResponse(200, httpResponseHeaders, "{}")
-            coEvery { http.send(httpRequest) } returns httpResponseWithoutClientId
-
-            var capturedError: PayPalSDKError? = null
-            try {
-                sut.fetchCachedOrRemoteClientID()
-            } catch (e: PayPalSDKError) {
-                capturedError = e
-            }
-            assertEquals(Code.DATA_PARSING_ERROR.ordinal, capturedError?.code)
-            assertEquals("sample-correlation-id", capturedError?.correlationID)
-        }
-
-    @Test
-    fun `fetchCachedOrRemoteClientID() throws server response error when http response is unsuccessful`() =
-        runTest {
-
-            every {
-                httpRequestFactory.createHttpRequestFromAPIRequest(any(), any())
-            } returns httpRequest
-
-            val httpResponseWithoutClientId = HttpResponse(500, httpResponseHeaders)
-            coEvery { http.send(httpRequest) } returns httpResponseWithoutClientId
-
-            var capturedError: PayPalSDKError? = null
-            try {
-                sut.fetchCachedOrRemoteClientID()
-            } catch (e: PayPalSDKError) {
-                capturedError = e
-            }
-            assertEquals(Code.SERVER_RESPONSE_ERROR.ordinal, capturedError?.code)
-            assertEquals("sample-correlation-id", capturedError?.correlationID)
-        }
-
-    @Test
-    fun `sendAnalyticsEvent() event delegates it to analytics service`() = runTest {
-        API.clientIDCache.put("fake-access-token", "fake-client-id")
-
-        sut.sendAnalyticsEvent("sample.event.name")
-        coVerify(exactly = 1) {
-            analyticsService.sendAnalyticsEvent("sample.event.name", "fake-client-id")
-        }
     }
 }
